# desktop/src-tauri/src/handlers/tag.rs
use crate::db::models::Tag;
use crate::db::{connection, DbState, TagRepository};
use crate::error::{AppError, Result};
use crate::handlers::common::{PaginatedTags, PaginationResponse};
use crate::utils::log_create;
use axum::{
    extract::{Query, State},
    http::StatusCode,
    response::{IntoResponse, Json},
};
use serde::Deserialize;
use utoipa::ToSchema;

#[derive(Deserialize, ToSchema)]
pub struct CreateTagRequest {
    pub name: String,
}

#[derive(Deserialize, ToSchema)]
pub struct BulkCreateTagsRequest {
    pub tags: Vec<CreateTagRequest>,
}

/// Get all tags
#[utoipa::path(
    get,
    path = "/v1/tags",
    tag = "Tags",
    params(
        ("limit" = Option<u32>, Query, description = "Number of tags per page (max 1000)"),
        ("cursor" = Option<String>, Query, description = "Cursor for pagination")
    ),
    responses(
        (status = 200, description = "Paginated list of tags", body = PaginatedTags),
        (status = 500, description = "Internal server error")
    )
)]
pub async fn get_all_tags(
    State(state): State<DbState>,
    Query(params): Query<crate::commands::params::PaginationQueryParams>,
) -> Result<impl IntoResponse> {
    let repo = TagRepository::new(connection::get_database(&state));
    let (tags, next_cursor, has_more) = repo
        .find_all(params.normalize_limit(), params.cursor)
        .await?;
    Ok((
        StatusCode::OK,
        Json(PaginationResponse::new(tags, next_cursor, has_more)),
    ))
}

/// Create a new tag
#[utoipa::path(
    post,
    path = "/v1/tags",
    tag = "Tags",
    request_body = CreateTagRequest,
    responses(
        (status = 200, description = "Created tag", body = Tag),
        (status = 400, description = "Bad request"),
        (status = 500, description = "Internal server error")
    )
)]
pub async fn create_tag(
    State(state): State<DbState>,
    Json(payload): Json<CreateTagRequest>,
) -> Result<impl IntoResponse> {
    if payload.name.is_empty() {
        return Err(AppError::BadRequest("Tag name cannot be empty".to_string()));
    }

    let db = connection::get_database(&state);
    let repo = TagRepository::new(db.clone());
    let tag = repo.create(payload.name).await?;
    
    // Log activity
    if let Err(e) = log_create(db, "tag".to_string(), tag.id.clone()).await {
        tracing::warn!("Failed to log tag creation activity: {}", e);
    }
    
    Ok((StatusCode::OK, Json(tag)))
}

/// Bulk create tags
#[utoipa::path(
    post,
    path = "/v1/tags/bulk-create",
    tag = "Tags",
    request_body = Vec<CreateTagRequest>,
    responses(
        (status = 200, description = "Created tags", body = Vec<Tag>),
        (status = 400, description = "Bad request"),
        (status = 500, description = "Internal server error")
    )
)]
pub async fn bulk_create_tags(
    State(state): State<DbState>,
    Json(payload): Json<Vec<CreateTagRequest>>,
) -> Result<impl IntoResponse> {
    let db = connection::get_database(&state);
    let repo = TagRepository::new(db.clone());
    let names: Vec<String> = payload.into_iter().map(|t| t.name).collect();
    let tags = repo.bulk_create(names).await?;
    
    // Log activities for each created tag
    for tag in &tags {
        if let Err(e) = log_create(db.clone(), "tag".to_string(), tag.id.clone()).await {
            tracing::warn!("Failed to log tag creation activity: {}", e);
        }
    }
    
    Ok((StatusCode::OK, Json(tags)))
}